pyproject.toml
[project]
name = "weathertools"
version = "0.1.0"
description = "中央氣象署天氣資料與監視器視覺分析工具"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.1.0",
    "playwright>=1.58.0",
]

[project.optional-dependencies]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# 新增以下這段
[tool.hatch.build.targets.wheel]
packages = ["."]
